[thresholding] remove bug affecting input width in top level wrapper
The C_BITS parameter is calculating the correct width needed for the top level wrapper for the thresholding binary search IP.
However, the parameter is not 'synthesizing' correctly and does not update the width for the affected s_axilite_AWADDR signal.
This results in the MSBs of the input signal being truncated. These missing bits affected addressing when writing weights into the core.
Weights were written to the incorrect addresses in the core causing incorrect thresholding to occur.

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>