[thresholding] adjust i/o port names on thresholding RTL wrapper
Originally s_axis and m_axis port names on the thresholding RTL wrapper could synthesise, but did not adhere to the FINN i/o signal naming convention.
The FINN compiler would not recognise the IP being synthesised and would rely on the IP having the correct IP/signal wiring steps in place.
The FINN compiler did not recognise s_axis/m_axis signal naming and therefore did not automatically set the clock frequency of the IP to match the rest of the network.
This required a Tcl command to set the clock frequency of the IP, as well a user-configurable attribute to set the clock frequency for ease-of-use.

It turns out that this actually reduces user ease-of-use. Having the compiler take care of the clock signalling is preferred. To do this, the s_axis/m_axis
signals are renamed to in0_V/out_V, as the compiler expects, and this extra 'user configurability' can therefore be removed.

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>