[thresholding]: remove reset that erases the 0th stage threshold value
There is a corner case where the number of channels configured for the thresholding core is 1 and during programming the stage 0 threshold parameter.
For each other stage in this case (and all other cases) the threshold parameters are non-volatile. When a reset happens after programming the threshold parameters, all would still be intact except for the 0th stage threshold value.

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>