[thresholding] only adjust MSB thresholding addressing bits when channel fold factor is present
In the case where channel fold is not present (i.e. CF is 0), we saw incorrect threshold address programming.
Without this commit and when no channel folding is present, this if statement is always stepped through and
was damaging LSBs of the Addr signal, causing incorrect threshold address programming for a PE core.

Although the logic in the if statement looks correct (i.e. programming 0 bits (clog2(CF) => 0)) and should not
harm the Addr signal, it's best to avoid stepping through a case that does not exist (i.e., there is no channel
folding and each channel has its own PE; therefore no extra bits needed to program multiple channel thresholds
into a single PE core).

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>