Remove postFailure() and postSuccess() functions.
This is an attempt to reduce the method count used in the pipeline as the current size causes the "groovyjarjarasm.asm.MethodTooLargeException: Method too large"
error. As a result the pipeline does not run at all.

This is a well known limitation. Removing unneccessary functions shrinks the method count and allows the Jenkinsfile to run.

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>