[Attention] Prevent absorbing thresholds into MVAU after forking matmul
Absorbing only the first of the consumer branches detaches the rest of
the graph. Absorbing all branches is not possible, as these might
actually be different thresholds or even other types of ops. So in these
cases now a matmul-only MVAU followed by standalone thresholding will be
generated, keeping the graph intact.