[Attention] Introduce ReplicateStream custom operation
This is intended to replicate the stream feeding the query, key and
value projections, thus producing three replicas of the stream. However,
it seems like DuplicateStreams_Batch indeed already supports an
arbitrary number of replicas, contrary to what the name and
corresponding transformation suggest... Then this new operation might be
redundant.