[Attention] Add graph transformations for handling attention operators
These are the new Infer* transformations for detecting the scaled
dot-product attention and multi-head attention patterns to convert these
to the hardware custom operations, as well as the InferReplicateStream
transformation, corresponding to the new ReplicateStream operation.

This also adds a few reordering transformations for optimizing the
attention pattern and eventually unrolling the attention heads within
the ONNX graph to achieve a fully-parallel hardware implementation of
multi-head attention, which might look like streamline transformations,
but must be applied when/after converting the attention pattern to
hardware operators.