[Squeeze] Introduce Squeeze and Unsqueeze hardware operators
This includes HWCustomOp and HLSBackend specializations of the operators
aiming for full ONNX compliance. Adds infrastructure for converting the
standard ONNX version of the operators to the FINN dialect, which mostly
means transplanting the node into the FINN domain and setting a few type
and shape attributes. Adds unit tests in Python, C++ and RTL simulation
as well as a simple integration test starting from PyTorch model export.