[Docker] Derive FINN_DOCKER_TAG from the SCRIPTPATH
Previously the FINN_DOCKER_TAG has been derived by "git describe" in the
current working directory, which fails if this is not the base of the
FINN repository. As we know the SCRIPTPATH, which always resides in the
base of the FINN repository, we can nagivate there, derive the docker
tag and navigate back to the working directory.